Add () or ( in completionText
